+++ src/main/java/ru/erulaev/restaurantvoting/common/model/BaseEntity.java
@@ -18,7 +18,7 @@ public abstract class BaseEntity implements HasId {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Schema(accessMode = Schema.AccessMode.READ_ONLY) // https://stackoverflow.com/a/28025008/548473
-    protected Long id;
+    protected Integer id;
 
     // https://stackoverflow.com/questions/1638723
     @Override
